+++ test/react-native/features/fixtures/ios-module/Scenarios/UserNativeClientScenario.m
@@ -1,0 +1,10 @@
+#import "UserNativeClientScenario.h"
+
+@implementation UserNativeClientScenario
+
+- (void)run {
+  [Bugsnag setUser:@"123" withEmail:@"bug@sn.ag" andName:@"Bug Snag"];
+  @throw [[NSException alloc] initWithName:@"NSException" reason:@"UserNativeClientScenario" userInfo:nil];
+}
+
+@end